requirements.txt
Flask
face_recognition
numpy
requests
pillow
